+++ ChangeLog.markdown
@@ -11,5 +11,6 @@ PHPUnit 3.7.0
 * Implemented #377: Show messages and stracktraces in JSON output for skipped and incomplete tests.
 * Fixed #440: Possible crash when using `--process-isolation` with PHP 5.3 and `detect_unicode=on`.
 * It is possible again to expect the generic `Exception` class.
+* Removed deprecated `--skeleton-class` and `--skeleton-test` switches. The functionality is now provided by the `phpunit-skel` command of the `PHPUnit_SkeletonGenerator` package.
 * Removed deprecated `PHPUnit_Extensions_OutputTestCase` class.
 * PHPUnit 3.7 is only supported on PHP 5.3.9 (or later) and PHP 5.4.0 (or later) is highly recommended.